requirements.txt
transformer
paddleocr
streamlit
torch
